+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/models/shared/journalline.py
@@ -6,6 +6,7 @@ from ..shared import accountref as shared_accountref
 from ..shared import recordref as shared_recordref
 from codatsynccommerce import utils
 from dataclasses_json import Undefined, dataclass_json
+from decimal import Decimal
 from typing import Optional
 
 
@@ -23,7 +24,7 @@ class JournalLineTracking:
 
 @dataclasses.dataclass
 class JournalLine:
-    net_amount: float = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('netAmount') }})
+    net_amount: Decimal = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('netAmount'), 'encoder': utils.decimalencoder(False, False), 'decoder': utils.decimaldecoder }})
     r"""Amount for the journal line. Debit entries are considered positive, and credit entries are considered negative."""
     account_ref: Optional[shared_accountref.AccountRef] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('accountRef'), 'exclude': lambda f: f is None }})
     r"""Data types that reference an account, for example bill and invoice line items, use an accountRef that includes the ID and name of the linked account."""