+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/sdk.py
@@ -31,11 +31,13 @@ from codatsynccommerce import utils
 from codatsynccommerce.models import shared
 
 class CodatSyncCommerce:
-    r"""Sync for Commerce (v1): The API for Sync for Commerce.
+    r"""Sync for Commerce (v1): The API for Sync for Commerce V1.
 
-    Sync for Commerce is an API and a set of supporting tools built to enable e-commerce and point of sale platforms to provide high-quality integrations with numerous accounting platform through standardized API, seamlessly transforming business sale's data into accounting artefacts.
+    Sync for Commerce automatically replicates and reconciles sales data from a merchant’s source PoS, Payments, and eCommerce systems into their accounting software. This eliminates manual processing by merchants and transforms their ability to run and grow their business.
 
     [Read More...](https://docs.codat.io/commerce/overview)
+
+    Not seeing what you expect? [See the main Sync for Commerce API](https://docs.codat.io/sync-for-commerce-api).
     """
     accounting_accounts: AccountingAccounts
     r"""Accounts"""
@@ -95,7 +97,8 @@ class CodatSyncCommerce:
                  server_idx: int = None,
                  server_url: str = None,
                  url_params: dict[str, str] = None,
-                 client: requests_http.Session = None
+                 client: requests_http.Session = None,
+                 retry_config: utils.RetryConfig = None
                  ) -> None:
         """Instantiates the SDK configuring it with the provided parameters.
         
@@ -108,7 +111,9 @@ class CodatSyncCommerce:
         :param url_params: Parameters to optionally template the server URL with
         :type url_params: dict[str, str]
         :param client: The requests.Session HTTP client to use for all operations
-        :type client: requests_http.Session        
+        :type client: requests_http.Session
+        :param retry_config: The utils.RetryConfig to use globally
+        :type retry_config: utils.RetryConfig
         """
         if client is None:
             client = requests_http.Session()
@@ -119,7 +124,7 @@ class CodatSyncCommerce:
             if url_params is not None:
                 server_url = utils.template_url(server_url, url_params)
 
-        self.sdk_configuration = SDKConfiguration(client, security_client, server_url, server_idx)
+        self.sdk_configuration = SDKConfiguration(client, security_client, server_url, server_idx, retry_config=retry_config)
        
         self._init_sdks()